requirements_dev.txt
-e .
beautifulsoup4
build
nox
pytest-cov
recommonmark
twine
defusedxml
